+++ alfresco-health-processor-platform/src/main/amp/config/alfresco/subsystems/HealthProcessor/default/webscripts-context.xml
@@ -4,14 +4,8 @@
         xsi:schemaLocation="http://www.springframework.org/schema/beans
 		http://www.springframework.org/schema/beans/spring-beans.xsd">
 
-    <bean id="eu.xenit.alfresco.healthprocessor.webscripts.console.AdminConsoleWebScript"
-            class="eu.xenit.alfresco.healthprocessor.webscripts.console.AdminConsoleWebScript"
-            parent="webscript">
-        <constructor-arg ref="Admin-Console.ResponseViewRenderer"/>
-    </bean>
-
     <bean id="Admin-Console.ResponseViewRenderer"
-            class="eu.xenit.alfresco.healthprocessor.webscripts.console.ResponseViewRenderer"
+            class="eu.xenit.alfresco.healthprocessor.webscripts.console.ResponseViewRendererImpl"
             parent="webscript" autowire="byType">
         <property name="serviceRegistry" ref="ServiceRegistry"/>
     </bean>
